Improvements to process/threading code plus a few changes to the user API.
Improved reference counting on processes. Also track how many threads are
running in a process so that we can free up resources (address space,
handles) as soon as the last thread terminates. Previously this was only
done when a thread is actually destroyed, which could in fact be a long
time after it exits if anything has a handle open to it.

Note that kern_process_clone() is currently unimplemented, I've changed it
to basically work the same as fork does - while the previous behaviour was
easy to implemented in the kernel, it is a bit messy to deal with in user
mode when implementing fork because we have to map a temporary stack for
the clone just to longjmp off it back to the original stack.

+++ source/kernel/include/object.h
@@ -143,7 +143,7 @@ extern status_t object_handle_attach(object_handle_t *handle, handle_t *idp,
 extern status_t object_handle_detach(handle_t id);
 
 extern status_t handle_table_create(handle_table_t *parent, handle_t map[][2],
-	int count, handle_table_t **tablep);
+	ssize_t count, handle_table_t **tablep);
 extern handle_table_t *handle_table_clone(handle_table_t *src);
 extern void handle_table_destroy(handle_table_t *table);
